+++ main.go
@@ -4,27 +4,47 @@ import (
 	"corpos-christie/colors"
 	"corpos-christie/config"
 	"corpos-christie/core"
+	"corpos-christie/user"
 	"corpos-christie/utils"
 	"fmt"
 	"log"
 	"os"
 )
 
+// Configuration of the application
 var cfg *config.Config
 
 // Start tax calculator from input user
-func start(cfg *config.Config) bool {
-	fmt.Print("Enter your income (Revenu net imposable): ")
-	var input string = utils.ReadValue()
-	income, err := utils.ConvertStringToInt(input)
+func start(cfg *config.Config, user *user.User) bool {
+	// Ask income's user
+	_, err := user.AskIncome()
 	if err != nil {
-		log.Printf("Error: Tax income is not convertible in int, details: %v", err)
+		log.Printf("Error: asking income for user, details: %v", err)
+		return false
+	}
+
+	// Ask if user is in couple
+	_, err = user.AskIsInCouple()
+	if err != nil {
+		log.Printf("Error: asking is in couple for user, details: %v", err)
+		return false
+	}
+
+	// Ask if user hasChildren
+	_, err = user.AskHasChildren()
+	if err != nil {
+		log.Printf("Error: asking has children, details: %v", err)
 		return false
 	}
 
 	// calculate tax
-	result := core.Process(income, cfg)
-	fmt.Printf("Income:\t\t%v €\nTax:\t\t%v €\nRemainder:\t%v €\n", colors.Red(result.Income), colors.Red(result.Tax), colors.Red(result.Remainder))
+	result := core.Process(user, cfg)
+
+	user.Tax = result.Tax
+	user.Remainder = result.Remainder
+
+	// show user
+	user.Show()
 
 	return true
 }
@@ -57,9 +77,12 @@ func main() {
 	log.Printf("Project: %v", colors.Yellow(cfg.Name))
 	log.Printf("Version: %v", colors.Yellow(cfg.Version))
 
+	// Init user
+	var user *user.User = new(user.User)
+
 	var keep bool
 	for ok := true; ok; ok = keep {
-		status := start(cfg)
+		status := start(cfg, user)
 		if status {
 			log.Println("Core process successful")
 		} else {