+++ user/struct.go
@@ -1,0 +1,11 @@
+package user
+
+// Define a user
+type User struct {
+	Income     int
+	IsInCouple bool
+	Tax        float64
+	Remainder  float64
+	Children   int
+	Parts      float64
+}